% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cellPbnPlot}
\alias{cellPbnPlot}
\title{Plot a cell PBN}
\usage{
cellPbnPlot(ingres.object, cell.id)
}
\arguments{
\item{ingres.object}{An \code{\linkS4class{ingres}} object with a non-null
\code{single.cell.pbn} slot.}

\item{cell.id}{A string containing the id of the cell to be plotted.}
}
\value{
A \code{ggplot} object containing the cell PBN plot.
}
\description{
Plots the PBN produced by \code{\link{computePbnBySingleCell}} for a
given cell and returns it as a ggplot object
}
\examples{
createExampleIngresObjectWithViper
ing <- computePbnBySingleCell(ing)
cellPbnPlot(ing, "sample1@ACAGCTAAGATCCCGC-1")

}
