% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePbn.R
\name{computePbnByCluster}
\alias{computePbnByCluster}
\title{Compute a Probabilistic Boolean Network (PBN) for each cluster.}
\usage{
computePbnByCluster(ingres.object)
}
\arguments{
\item{ingres.object}{An \code{\linkS4class{ingres}} object with a non-null VIPER slot.}
}
\value{
An \code{\linkS4class{ingres}} object with the \code{cluster.pbn slot} filled
}
\description{
After running \code{\link{performViper}}, this can be run to produce a PBN with
the identity function probability for each node relative to the median VIPER
normalised enrichment score (NES) for each cluster.
}
