% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePbn.R
\name{computePbnBySingleCell}
\alias{computePbnBySingleCell}
\title{Compute a Probabilistic Boolean Network (PBN) for each cell.}
\usage{
computePbnBySingleCell(ingres.object)
}
\arguments{
\item{ingres.object}{An \linkS4class{ingres} object with a non-null VIPER slot.}
}
\value{
An \linkS4class{ingres} object with the single.cell.pbn slot filled
}
\description{
After running \code{\link{performViper}}, this can be run to produce a PBN with
the identity function probability for each node relative to the VIPER
normalised enrichment score (NES) for every cell.
}
