% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{createExampleIngresObject}
\alias{createExampleIngresObject}
\alias{createExampleIngresObjectWithViper}
\title{Create a toy \code{\linkS4class{ingres}} object to use in examples and tests.}
\usage{
createExampleIngresObject()

createExampleIngresObjectWithViper()
}
\value{
An \code{\linkS4class{ingres}} object
}
\description{
\code{createExampleIngresObjectWithViper additionally
fills the viper slot with a dummy viper object}
}
\examples{
createExampleIngresObject()
createExampleIngresObjectWithViper()
}
