% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingressClass.R
\name{createIngresObject}
\alias{createIngresObject}
\title{Create a new \code{ingres} object.}
\usage{
createIngresObject(expression.matrix, idents, network.genes, network)
}
\arguments{
\item{expression.matrix}{A single-cell expression matrix, with cells in rows and genes in columns.}

\item{idents}{A data frame with a column for cell barcode and another column for the corresponding cluster or subpopulation.}

\item{network.genes}{A data frame with a column for node names and another column for the corresponding entrez IDs.}

\item{network}{A object of class \code{tbl_graph} representing our network of interest.}
}
\value{
An \code{ingres} object.
}
\description{
Create a new \code{ingres} object.
}
