% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netwokHelpers.R
\name{createNetworkGenesTemplate}
\alias{createNetworkGenesTemplate}
\title{Create a network genes data frame. Optionally store it as csv
and open it for editing
To create an ingres object, a data frame with the network nodes and
the corresponding gene symbols must be provided. This function simplifies
the process. If the gene nodes are correct gene symbols, then modification
is not needed and the returned data frame can be directly passed to
the ingres constructors.}
\usage{
createNetworkGenesTemplate(network, dir = getwd(), store = TRUE, modify = TRUE)
}
\arguments{
\item{network}{A tidygraph network.}

\item{dir}{The directory where the csv will be stored, if applicable.}

\item{store}{If true, store the data frame as a csv}

\item{modify}{If true, and store is also true,
open it to be modified by the user}
}
\value{
The template data frame.
}
\description{
Create a network genes data frame. Optionally store it as csv
and open it for editing
To create an ingres object, a data frame with the network nodes and
the corresponding gene symbols must be provided. This function simplifies
the process. If the gene nodes are correct gene symbols, then modification
is not needed and the returned data frame can be directly passed to
the ingres constructors.
}
\examples{
tmp = tempdir()
createNetworkGenesTemplate(network, dir = tmp)
file.remove(paste(tmp, "/networkGenes.csv")) # cleanup

}
