% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingresClass.R
\name{getExpressionMatrix}
\alias{getExpressionMatrix}
\title{Get the Expression Matrix from an \code{\linkS4class{ingres}} object}
\usage{
getExpressionMatrix(ingres.object)
}
\arguments{
\item{ingres.object}{an \code{\linkS4class{ingres}} object}
}
\value{
An expression matrix
}
\description{
Get the Expression Matrix from an \code{\linkS4class{ingres}} object
}
\examples{
ing <- createExampleIngresObject()
head(GetExpressionMatrix(ing))
}
