% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingresClass.R
\name{getIdents}
\alias{getIdents}
\title{Get the idents from an \code{\linkS4class{ingres}} object}
\usage{
getIdents(ingres.object)
}
\arguments{
\item{ingres.object}{an \code{\linkS4class{ingres}} object}
}
\value{
A data frame with a column for cell barcode
and another column for the corresponding cluster or subpopulation.
}
\description{
Get the idents from an \code{\linkS4class{ingres}} object
}
\examples{
ing <- createExampleIngresObject()
head(GetIdents(ing))
}
