% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingresClass.R
\name{getNetworkGenes}
\alias{getNetworkGenes}
\title{Get the network genes from an \code{\linkS4class{ingres}} object}
\usage{
getNetworkGenes(ingres.object)
}
\arguments{
\item{ingres.object}{an \code{\linkS4class{ingres}} object}
}
\value{
A data frame with a column for node names
and another column for the corresponding entrez IDs.
}
\description{
Get the network genes from an \code{\linkS4class{ingres}} object
}
\examples{
ing <- createExampleIngresObject()
head(tetNetworkGenes(ing))
}
