% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingresClass.R
\name{getSingleCellPbns}
\alias{getSingleCellPbns}
\title{Get the single-cell PBNs from an \code{\linkS4class{ingres}} object}
\usage{
getSingleCellPbns(ingres.object)
}
\arguments{
\item{ingres.object}{an \code{\linkS4class{ingres}} object}
}
\value{
A data frame containing the generated single-cell PBNs
}
\description{
Get the single-cell PBNs from an \code{\linkS4class{ingres}} object
}
\examples{
ing <- createExampleIngresObjectWithViper()
ing <- ing \%>\% computePbnBySingleCell()
head(getSingleCellPbns(ing))
}
