% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingresClass.R
\name{getViperResults}
\alias{getViperResults}
\title{Get the viper results from an \code{\linkS4class{ingres}} object}
\usage{
getViperResults(ingres.object)
}
\arguments{
\item{ingres.object}{an \code{\linkS4class{ingres}} object}
}
\value{
A data frame containing viper results
}
\description{
Get the viper results from an \code{\linkS4class{ingres}} object
}
\examples{
ing <- createExampleIngresObjectWithViper()
head(getViperResults(ing))
}
