% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netwokHelpers.R
\name{graphmlAsTidy}
\alias{graphmlAsTidy}
\title{Convert GraphML format to tidygraph}
\usage{
graphmlAsTidy(filename)
}
\arguments{
\item{filename}{The path to the the GraphML file.}
}
\value{
A tidygraph object
}
\description{
In order to be used by ingres,
GraphML files have to be converted to tidygraph format. This helper does that.
}
\examples{
\dontrun{
graphmlAsTidy("network.graphml")
}

# With the example file:
graphmlAsTidy(system.file("extdata", "network.graphml", package = "ingres"))
}
