% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingresClass.R
\docType{class}
\name{ingres-class}
\alias{ingres-class}
\alias{ingres}
\alias{ingresClass}
\alias{ingresObject}
\title{The ingres class.}
\value{
An \code{ingres} object.
}
\description{
The ingres class.
}
\section{Slots}{

\describe{
\item{\code{expression}}{matrix.}

\item{\code{idents}}{data.frame.}

\item{\code{network.genes}}{data.frame.}

\item{\code{viper}}{data.frame.}

\item{\code{cluster.pbn}}{data.frame.}

\item{\code{single.cell.pbn}}{data.frame.}

\item{\code{network}}{\code{tbl_graph}.}
}}

