% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network}
\alias{network}
\title{An example network,
in the format required by the package}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 58.
}
\usage{
network
}
\description{
An example network,
in the format required by the package
}
\keyword{datasets}
