% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network_genes}
\alias{network_genes}
\title{An example dataframe with the genes contained in the network,
in the format required by the package}
\format{
An object of class \code{data.frame} with 28 rows and 2 columns.
}
\usage{
network_genes
}
\description{
An example dataframe with the genes contained in the network,
in the format required by the package
}
\keyword{datasets}
