% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performViper.R
\name{performViper}
\alias{performViper}
\title{Perform VIPER}
\usage{
performViper(ingres.object, regulon)
}
\arguments{
\item{ingres.object}{An ingres object.}

\item{regulon}{Object of class regulon or list of objects of class regulon for metaVIPER analysis}
}
\value{
An ingres object with the result of the VIPER inference in the \code{viper} slot
}
\description{
Performs VIPER analysis on a RNA-seq dataset contained by an ingres object.
This function is merely a wrapper around the \code{\link[viper:viper]{original VIPER function}},
designed to suit the ingres workflow.
}
