% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSelectedCell.R
\name{plotSelectedCell}
\alias{plotSelectedCell}
\title{Plot the PBN of a cell selected in an interactive plot.}
\usage{
plotSelectedCell(ingres.object, seurat.object)
}
\arguments{
\item{ingres.object}{An \code{\linkS4class{ingres}} object with a non-null
\code{single.cell.pbn} slot.}

\item{seurat.object}{A \code{\link[SeuratObject:Seurat-class]{Seurat object}}
created with the same dataset as the ingres object. Ideally, this object
was used to create the ingres object through
\code{\link{createIngresObjectFromSeurat}}.}
}
\description{
Launches a Shiny app with an interactive tSNE plot from a Seurat object. When
a cell is clicked, the PBN for that cell is plotted side by side
with the tSNE plot.
Internally, it uses \code{\link{cellPbnPlot}} on the selected cell.
}
\examples{
\dontrun{
ing = createIngresObjectFromSeurat(
  small_blca_wang, "RNA", "data", network_genes, network
)
ing@viper = viper_results
ing = computePbnBySingleCell(ing)
plotSelectedCell(ing, small_blca_wang)
}

}
