% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceNetwork.R
\name{produceNetworkForCluster}
\alias{produceNetworkForCluster}
\title{Produce a tidygraph network for a given cluster.}
\usage{
produceNetworkForCluster(ingres.object, cluster.id)
}
\arguments{
\item{ingres.object}{The ingres object containing the relevant data.}

\item{cluster.id}{The id of the cluster for which the network is to be produced.}
}
\value{
A tidygraph object containing the network for the given cluster.
}
\description{
Produce a tidygraph network for a given cluster.
}
