% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_blca_wang}
\alias{small_blca_wang}
\title{An example small Seurat object, using dataset GSE130001 from
Wang et al. Genome Med. 2020 Dec;12(1):24.}
\format{
An object of class \code{Seurat} with 16314 rows and 90 columns.
}
\usage{
small_blca_wang
}
\description{
An example small Seurat object, using dataset GSE130001 from
Wang et al. Genome Med. 2020 Dec;12(1):24.
}
\keyword{datasets}
