% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viper_results}
\alias{viper_results}
\title{The results of running \code{\link[=performViper]{performViper()}} on an ingres object created with
\code{small_blca_wang}, intended to speed up vignettes and examples.}
\format{
An object of class \code{data.frame} with 90 rows and 4945 columns.
}
\usage{
viper_results
}
\description{
The results of running \code{\link[=performViper]{performViper()}} on an ingres object created with
\code{small_blca_wang}, intended to speed up vignettes and examples.
}
\keyword{datasets}
