% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viperResults}
\alias{viperResults}
\title{An example dataframe with VIPER results, meant to speed up testing and demonstrations}
\format{
An object of class \code{data.frame} with 3483 rows and 81 columns.
}
\usage{
viperResults
}
\description{
An example dataframe with VIPER results, meant to speed up testing and demonstrations
}
\keyword{datasets}
