% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{clusterPbnPlot}
\alias{clusterPbnPlot}
\title{Plot a cluster PBN}
\usage{
clusterPbnPlot(ingres.object, cluster.id)
}
\arguments{
\item{ingres.object}{An \code{\linkS4class{ingres}} object with a non-null
\code{cluster.pbn} slot.}

\item{cluster.id}{A string containing the id of the cluster to be plotted.}
}
\value{
A \code{ggplot} object containing the cluster PBN plot.
}
\description{
Plots the PBN produced by \code{\link{computePbnByCluster}} for a given
cluster and returns it as a ggplot object
}
\examples{
# Create an ingres object with viper slot
ing = createIngresObjectFromSeurat(
  small_blca_wang, "RNA", "data", network_genes, network
)
ing@viper = viper_results

ing = computePbnByCluster(ing)
clusterPbnPlot(ing, "1")

}
