% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePbn.R
\name{computePbnByCluster}
\alias{computePbnByCluster}
\title{Compute a Probabilistic Boolean Network (PBN) for each cluster.}
\usage{
computePbnByCluster(ingres.object, range = c(-1, 1))
}
\arguments{
\item{ingres.object}{An \code{\linkS4class{ingres}} object with a non-null VIPER slot.}

\item{range}{A vector representing the range to which the NES are to be calculated.
\code{range[1]+range[2] == 0} should be TRUE. Defaults to [-1, 1]}
}
\value{
An \code{\linkS4class{ingres}} object with the \code{cluster.pbn slot} filled
}
\description{
After running \code{\link{performViper}}, this can be run to produce a PBN with
the identity function probability for each node relative to the median VIPER
normalised enrichment score (NES) for each cluster.
}
