% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingres-package.R
\docType{package}
\name{ingres-package}
\alias{ingres-package}
\alias{_PACKAGE}
\title{ingres: Infer Gene Probabilistic Boolean Networks from Single-Cell Data}
\description{
Given a gene regulatory boolean network and a RNA-seq dataset, this package computes protein activity normalised enrichment scores using 'VIPER', and then produces a probabilistic network using the scores as probabilities for fixed node activation or deactivation, in addition to the original Boolean functions. For more information, refer to the preprint: Victori and Buffa (2022) <doi:10.1101/2022.09.04.506528>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CBigOxf/ingres}
  \item \url{https://cbigoxf.github.io/ingres/}
  \item Report bugs at \url{https://github.com/CBigOxf/ingres/issues}
}

}
\author{
\strong{Maintainer}: Pedro Victori \email{pedroluisvictori@gmail.com} (\href{https://orcid.org/0000-0001-6321-5584}{ORCID})

Authors:
\itemize{
  \item Francesca M. Buffa \email{francesca.buffa@unibocconi.it} (\href{https://orcid.org/0000-0003-0409-406X}{ORCID})
}

}
\keyword{internal}
