% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performViper.R
\name{performViper}
\alias{performViper}
\title{Perform VIPER}
\usage{
performViper(ingres.object, regulon, verbose = FALSE)
}
\arguments{
\item{ingres.object}{An ingres object.}

\item{regulon}{Object of class regulon or list of objects of class regulon for metaVIPER analysis}

\item{verbose}{Logical, whether progression messages should be printed in the terminal}
}
\value{
An ingres object with the result of the VIPER inference in
the \code{viper} slot
}
\description{
Performs VIPER analysis on a RNA-seq dataset contained by an ingres object.
This function is merely a wrapper around the
\code{\link[viper:viper]{original VIPER function}},
designed to suit the ingres workflow and properly prepare the input data.
}
\examples{
\dontshow{if (require("aracne.networks")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Subset regulon to speed up example run
regulon = aracne.networks::regulonblca[1:100]
ing = createIngresObjectFromSeurat(
  small_blca_wang, "RNA", "data", network_genes, network
)
performViper(ing, regulon)
\dontshow{\}) # examplesIf}
}
