% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netwokHelpers.R
\name{produceBoolnetNetwork}
\alias{produceBoolnetNetwork}
\title{Convert an ingres network into a BoolNet one.}
\usage{
produceBoolnetNetwork(network)
}
\arguments{
\item{network}{The network to be converted, created
by \code{\link[=produceNetworkForCluster]{produceNetworkForCluster()}} or \code{\link[=produceNetworkForCell]{produceNetworkForCell()}}}
}
\value{
A BoolNet object.
}
\description{
Convert an ingres network into a BoolNet one.
}
