% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceNetwork.R
\name{produceBoolnetNetwork}
\alias{produceBoolnetNetwork}
\title{Converts a tidygraph network into a BoolNet one.}
\usage{
produceBoolnetNetwork(tidy.network)
}
\arguments{
\item{tidy.network}{The network to be converted.}
}
\value{
A BoolNet object.
}
\description{
Converts a tidygraph network into a BoolNet one.
}
