% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceNetwork.R
\name{produceNetworkForCell}
\alias{produceNetworkForCell}
\title{Produce a tidygraph network for a given cell}
\usage{
produceNetworkForCell(ingres.object, cell.id)
}
\arguments{
\item{ingres.object}{The ingres object containing the relevant data.}

\item{cell.id}{The id of the cell for which the network is to be produced.}
}
\value{
A tidygraph object containing the network for the given cell.
}
\description{
Produce a tidygraph network for a given cell
}
\examples{
# Create an ingres object with viper slot
ing = createIngresObjectFromSeurat(
  small_blca_wang, "RNA", "data", network_genes, network
)
ing@viper = viper_results

ing = computePbnBySingleCell(ing)

network = produceNetworkForCell(ing, "sample1@ACAGCTAAGATCCCGC-1")
print(network)

}
